Streamline our middleware into one file.
